adress comments
