+++ libs/coin-modules/coin-internet_computer/src/common-logic/utils.ts
@@ -6,17 +6,17 @@ const validBase64RegExp = new RegExp(
   /^(?:[A-Za-z\d+/]{4})*(?:[A-Za-z\d+/]{3}=|[A-Za-z\d+/]{2}==)?$/,
 );
 
-export const isNoErrorReturnCode = (code: number): boolean => code === 0x9000;
+// export const isNoErrorReturnCode = (code: number): boolean => code === 0x9000;
 
-export const getPath = (path: string): string =>
-  path && path.substr(0, 2) !== "m/" ? `m/${path}` : path;
+// export const getPath = (path: string): string =>
+//   path && path.substr(0, 2) !== "m/" ? `m/${path}` : path;
 
 export const isValidHex = (msg: string): boolean => validHexRegExp.test(msg);
 export const isValidBase64 = (msg: string): boolean => validBase64RegExp.test(msg);
 
-export const isError = (r: { returnCode: number; errorMessage?: string }): void => {
-  if (!isNoErrorReturnCode(r.returnCode)) throw new Error(`${r.returnCode} - ${r.errorMessage}`);
-};
+// export const isError = (r: { returnCode: number; errorMessage?: string }): void => {
+//   if (!isNoErrorReturnCode(r.returnCode)) throw new Error(`${r.returnCode} - ${r.errorMessage}`);
+// };
 
 export const methodToString = (method: number): string => {
   switch (method) {
