+++ libs/ledger-live-common/src/families/internet_computer/common.ts
@@ -1,0 +1,8 @@
+export const isNoErrorReturnCode = (code: number): boolean => code === 0x9000;
+
+export const getPath = (path: string): string =>
+  path && path.substr(0, 2) !== "m/" ? `m/${path}` : path;
+
+export const isError = (r: { returnCode: number; errorMessage?: string }): void => {
+  if (!isNoErrorReturnCode(r.returnCode)) throw new Error(`${r.returnCode} - ${r.errorMessage}`);
+};
